# Speed up bar states bounds calculation of entropies