This PR speeds up the calculation of bounds for mathematical and specific entropy.
Before, we looped through every node, look at the bar states on the right, top, left, bottom and calculated the entropy every time. This resulted in doubled entropy calculation.

Now, we just look at the bar states on the left and bottom, calculate the entropy and save it for the node itself and the respective neighbor. In the end, we add the calculation at the interface/boundary far right and top.

The new implementation only takes about 70% of the time the old one took.
Before:
`calc_variable_bounds!         4.34k    992ms   20.4%   229μs    577KiB   19.9%     136B`
Now:
`calc_variable_bounds!         4.34k    704ms   15.5%   162μs    577KiB   19.9%     136B`

This is not necessary for the calculation of the bounds for the local limiting, since there we can directly use the bar states without further calculations.