@bennibolm 2023-11-20
I think this PR is directly mergeable. Of course, I checked that we get the same results as before and (hopefully) all tests pass.
Could you take a quick look anyway?@amrueda